MELT: Remove unnecessary stuffs
