+++ src/helm/benchmark/presentation/run_entries_melt.conf
@@ -111,40 +111,6 @@ entries: [
   {description: "melt_math:model=text_code,subject=counting_and_probability,level=5,use_chain_of_thought=True", priority: 2}
   {description: "melt_math:model=text_code,subject=precalculus,level=5,use_chain_of_thought=True", priority: 2}
 
-  # Text Classification
-  {description: "melt_text_classification_vsmec:model=text,fewshot=True", priority: 1}
-  {description: "melt_text_classification_phoatis:model=text,fewshot=True", priority: 1}
-
-  {description: "melt_text_classification_vsmec:model=text,fewshot=False", priority: 2}
-  {description: "melt_text_classification_phoatis:model=text,fewshot=False", priority: 2}
-
-  # Sentiment Analysis
-  {description: "melt_sentiment_analysis_vlsp:model=text,fewshot=True", priority: 1}
-  {description: "melt_sentiment_analysis_vsfc:model=text,fewshot=True", priority: 1}
-
-  {description: "melt_sentiment_analysis_vlsp:model=text,fewshot=False", priority: 2}
-  {description: "melt_sentiment_analysis_vsfc:model=text,fewshot=False", priority: 2}
-
-  # Translation
-  {description: "melt_translation_opus100:model=text,language_pair=en-vi,fewshot=True", priority: 1},
-  {description: "melt_translation_phomt:model=text,language_pair=en-vi,fewshot=True", priority: 1},
-
-  {description: "melt_translation_opus100:model=text,language_pair=vi-en,fewshot=True", priority: 2},
-  {description: "melt_translation_phomt:model=text,language_pair=vi-en,fewshot=True", priority: 2},
-
-  {description: "melt_translation_opus100:model=text,language_pair=en-vi,fewshot=False", priority: 3},
-  {description: "melt_translation_phomt:model=text,language_pair=en-vi,fewshot=False", priority: 3},
-
-  {description: "melt_translation_opus100:model=text,language_pair=vi-en,fewshot=False", priority: 4},
-  {description: "melt_translation_phomt:model=text,language_pair=vi-en,fewshot=False", priority: 4},
-
-  # Language Modeling
-  {description: "melt_lm_mask_filling_mlqa:model=full_functionality_text,fewshot=True", priority: 1}
-  {description: "melt_lm_spelling_correction_vsec:model=full_functionality_text,fewshot=True", priority: 1}
-
-  {description: "melt_lm_mask_filling_mlqa:model=full_functionality_text,fewshot=False", priority: 2}
-  {description: "melt_lm_spelling_correction_vsec:model=full_functionality_text,fewshot=False", priority: 2}
-
   # Knowledge
   {description: "melt_knowledge_zalo:model=text,data_augmentation=canonical,fewshot=True", priority: 1}
   {description: "melt_knowledge_vimmrc:model=text,randomize_order=False,data_augmentation=canonical,fewshot=True", priority: 1}
