+++ src/helm/benchmark/scenarios/melt_scenarios.py
@@ -1,12 +1,9 @@
 import os
 from abc import abstractmethod
-from typing import Any, Dict, List, Tuple, Optional
+from typing import Dict, List, Tuple, Optional
 
-import random
-from datasets import load_dataset, Dataset
-from helm.common.hierarchical_logger import hlog
+from datasets import load_dataset
 from helm.common.general import ensure_directory_exists
-from helm.common.hierarchical_logger import htrack_block
 from helm.benchmark.scenarios.scenario import (
     Scenario,
     Instance,
@@ -18,7 +15,6 @@ from helm.benchmark.scenarios.scenario import (
     PassageQuestionInput,
     Input,
     Output,
-    make_rank_tag,
 )
 from helm.benchmark.scenarios.math_scenario import get_answer
 
@@ -593,751 +589,6 @@ class MELTTextClassificationScenario(Scenario):
         return instances
 
 
-class MELTTextClassificationVSMECScenario(MELTTextClassificationScenario):
-    """
-    Scenario for the UIT-VSMEC dataset.
-    The UIT-VSMEC dataset is a Vietnamese emotion-labeled corpus consisting of
-    6,927 human-annotated sentences collected from social media, categorized
-    into six emotions: sadness, enjoyment, anger, disgust, fear, and surprise.
-    """
-
-    name = "melt_text_classification_vsmec"
-    description = "UIT-VSMEC dataset for emotion classification."
-    tags = ["text_classification"]
-
-    def __init__(self):
-        super().__init__(
-            dataset_name="ura-hcmut/UIT-VSMEC",
-            revision="ab642b189eff31fdb781cca7c4c34dee3ee0f1de",
-            text_key="Sentence",
-            label_key="Emotion",
-            splits={
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            },
-        )
-
-    def process_example(self, sample: dict) -> Tuple[str, List[str]]:
-        """
-        Given an sample from the dataset, create the input text and
-        list of answers for the instance.
-        """
-
-        return sample[self.text_key], [sample[self.label_key].lower()]
-
-
-class MELTTextClassificationPhoATISScenario(MELTTextClassificationScenario):
-    """
-    Scenario for the PhoATIS dataset.
-    The PhoATIS dataset is a Vietnamese benchmark for intent detection and slot filling,
-    consisting of 5,871 fluent utterances collected from task-oriented dialogue systems.
-    It was later extended with manual disfluency annotations to create a disfluent variant,
-    enabling research on the impact of disfluencies in spoken language understanding for Vietnamese.
-    """
-
-    name = "melt_text_classification_phoatis"
-    description = "PhoATIS dataset for intent detection of flight booking."
-    tags = ["text_classification"]
-
-    def __init__(self):
-        super().__init__(
-            dataset_name="ura-hcmut/PhoATIS",
-            revision="bd026c9b276d7fb083d19ec3d6870fca90e1834f",
-            text_key="text",
-            label_key="label",
-            splits={
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            },
-        )
-
-    def process_example(self, sample: dict) -> Tuple[str, List[str]]:
-        """
-        Given an sample from the dataset, create the input text and
-        list of answers for the instance.
-        """
-
-        return sample[self.text_key], sample[self.label_key].lower().split("#")
-
-
-class MELTTSentimentAnalysisVLSPScenario(MELTTextClassificationScenario):
-    """
-    Scenario for the VLSP 2016 sentiment analysis dataset.
-    The VLSP2016 dataset is a Vietnamese sentiment analysis corpus consisting of
-    short user-generated reviews from social media, each labeled with an overall
-    sentiment of positive, negative, or neutral. It was developed to support polarity
-    classification and benchmark Vietnamese sentiment analysis systems through the
-    VLSP 2016 evaluation campaign.
-    """
-
-    name = "melt_sentiment_analysis_vlsp"
-    description = "VLSP 2016 contains public comments from social media, used for sentiment analysis."
-    tags = ["sentiment_analysis"]
-
-    def __init__(self):
-        super().__init__(
-            dataset_name="ura-hcmut/vlsp2016",
-            revision="9531ec0ccabcafb7d51020fe69d8f9faebb91953",
-            text_key="Data",
-            label_key="Class",
-            splits={
-                TRAIN_SPLIT: "train",
-                TEST_SPLIT: "test",
-            },
-        )
-
-    def process_example(self, sample: dict) -> Tuple[str, List[str]]:
-        """
-        Given an sample from the dataset, create the input text and
-        list of answers for the instance.
-        """
-
-        return sample[self.text_key], [sample[self.label_key].lower()]
-
-
-class MELTTSentimentAnalysisVSFCScenario(MELTTextClassificationScenario):
-    """
-    Scenario for the UIT-VSFC dataset.
-    The UIT-VSFC dataset is a Vietnamese corpus of over 16,000 student feedback sentences,
-    annotated for both sentiment-based (positive, negative, neutral) and topic-based classifications.
-    It supports interdisciplinary research at the intersection of sentiment analysis and education,
-    with high inter-annotator agreement and strong baseline performance using a Maximum Entropy classifier.
-    """
-
-    name = "melt_sentiment_analysis_vsfc"
-    description = "UIT-VSFC dataset for analyzing sentiment of student feedback."
-    tags = ["sentiment_analysis"]
-
-    def __init__(self):
-        super().__init__(
-            dataset_name="ura-hcmut/UIT-VSFC",
-            revision="c572aed01a811a1dbc68e9aed9f9e684980a10a2",
-            text_key="text",
-            label_key="label",
-            splits={
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            },
-        )
-
-    def process_example(self, sample: dict) -> Tuple[str, List[str]]:
-        """
-        Given an sample from the dataset, create the input text and
-        list of answers for the instance.
-        """
-
-        return sample[self.text_key], [sample[self.label_key].lower()]
-
-
-class MELTTranslationScenario(Scenario):
-    """ """
-
-    name = "melt_translation"
-    description = "Machine Translation scenario."
-    tags = ["machine_translation"]
-
-    def __init__(
-        self,
-        dataset_name: str,
-        revision: str,
-        source_language: str,
-        target_language: str,
-        subset: Optional[str] = "",
-        splits: Optional[Dict[str, str]] = None,
-    ):
-        """Initializes the question answering scenario.
-
-        Args:
-            dataset_name: The name of the dataset.
-            revision: The revision of the dataset to use.
-            source_language: The source language to use.
-            target_language: The target language to use.
-            subset: The subset of the dataset to use. Defaults to "".
-            splits: The splits to use for the dataset. Defaults to None.
-        """
-        super().__init__()
-        self.MAX_TRAIN_INSTANCES = 20_000
-        valid_languages = set(["vi", "en"])
-        self.dataset_name = dataset_name
-        self.subset = subset
-        self.revision = revision
-        self.splits = splits
-        self.source_language = source_language
-        self.target_language = target_language
-        if self.source_language not in valid_languages or self.target_language not in valid_languages:
-            raise ValueError("Supported languages: vi, en.")
-        if self.source_language == self.target_language:
-            raise ValueError("The source language and the target language should be different.")
-        if self.source_language != "en" and self.target_language != "en":
-            raise ValueError("One of the languages should be English.")
-
-    def get_instances_for_splits(self, splits: Dict[str, str]) -> List[Instance]:
-        """
-        Helper for generating instances for a split.
-        Args:
-            splits (dict): Which splits to partition the data into.
-        Returns:
-            List[Instance]: Instances from the file for the specified split.
-        """
-        with htrack_block("Loading the HuggingFace dataset. The first time could take several minutes."):
-            hf_dataset: Any = load_dataset(
-                self.dataset_name,
-                self.subset,
-                revision=self.revision,
-                trust_remote_code=True,
-            )
-
-        instances: List[Instance] = []
-
-        for dataset_split_name, helm_split_name in splits.items():
-            if helm_split_name == TRAIN_SPLIT:
-                hf_dataset[dataset_split_name] = hf_dataset[dataset_split_name].shuffle(seed=42)[
-                    : self.MAX_TRAIN_INSTANCES
-                ]
-                hf_dataset[dataset_split_name] = Dataset.from_dict(hf_dataset[dataset_split_name])
-
-            for example in hf_dataset[dataset_split_name]:
-                source_sentence = example[self.source_language]
-                target_sentence = example[self.target_language]
-                instances.append(
-                    Instance(
-                        input=Input(text=source_sentence),
-                        references=[Reference(Output(text=target_sentence), tags=[CORRECT_TAG])],
-                        split=helm_split_name,
-                    )
-                )
-        return instances
-
-    def get_instances(self, output_path: str) -> List[Instance]:
-        if self.splits is None:
-            splits = {"train": TRAIN_SPLIT, "validation": VALID_SPLIT, "test": TEST_SPLIT}
-        else:
-            splits = {}
-            if "train" in self.splits:
-                splits[self.splits[TRAIN_SPLIT]] = TRAIN_SPLIT
-            if "validation" in self.splits:
-                splits[self.splits[VALID_SPLIT]] = VALID_SPLIT
-            if "test" in self.splits:
-                splits[self.splits[TEST_SPLIT]] = TEST_SPLIT
-
-        instances: List[Instance] = self.get_instances_for_splits(splits=splits)
-        return instances
-
-
-class MELTTranslationOPUS100Scenario(MELTTranslationScenario):
-    """
-    Scenario for the OPUS100 dataset.
-    """
-
-    name = "melt_translation_opus100"
-    description = "OPUS100 dataset for machine translation."
-    tags = ["machine_translation"]
-
-    def __init__(self, **kwargs):
-        super().__init__(
-            dataset_name="vietgpt/opus100_envi",
-            revision="45df06fb0b31edc882d7c8d34389261f995e5208",
-            splits={
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            },
-            **kwargs,
-        )
-
-
-class MELTTranslationPhoMTScenario(MELTTranslationScenario):
-    """
-    Scenario for the PhoMT dataset.
-    """
-
-    name = "melt_translation_phomt"
-    description = "PhoMT dataset for machine translation."
-    tags = ["machine_translation"]
-
-    def __init__(self, **kwargs):
-        super().__init__(
-            dataset_name="ura-hcmut/PhoMT",
-            revision="74386685db01dc038860ff0a90d9f5fbde284bf7",
-            splits={
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            },
-            **kwargs,
-        )
-
-
-class MELTLMMaskFillingScenario(Scenario):
-    """
-    Scenario for the MELT Masked Language Modeling dataset.
-    """
-
-    name = "melt_lm_mask_filling"
-    description = "Masked Language Modeling scenario."
-    tags = ["language_modeling", "mask_filling"]
-
-    def __init__(
-        self,
-        dataset_name: str,
-        revision: str,
-        masked_ratio: float = 0.1,
-        text_key: str = "text",
-        subset: Optional[str] = "",
-        splits: Optional[Dict[str, str]] = None,
-    ):
-        """Initializes the question answering scenario.
-
-        Args:
-            dataset_name: The name of the dataset.
-            revision: The revision of the dataset to use.
-            masked_ratio: The ratio of tokens to mask in the input text. Defaults to 0.1.
-            text_key: The key to use for the text in the dataset. Defaults to "text".
-            subset: The subset of the dataset to use. Defaults to "".
-            splits: The splits to use for the dataset. Defaults to None.
-        """
-        super().__init__()
-        self.dataset_name = dataset_name
-        self.subset = subset
-        self.masked_ratio = masked_ratio
-        self.text_key = text_key
-        self.revision = revision
-        self.splits = splits
-
-    def __mask_text__(self, text: str) -> str:
-        """
-        Mask a portion of the input text.
-        Args:
-            text (str): The input text to mask.
-        Returns:
-            str: The masked text.
-        """
-        tokens = text.split(" ")
-        num_tokens_to_mask = int(len(tokens) * self.masked_ratio)
-        indices_to_mask = random.sample(range(len(tokens)), num_tokens_to_mask)
-        for index in indices_to_mask:
-            tokens[index] = "[MASK]"
-        return " ".join(tokens)
-
-    def get_instances_for_splits(self, splits: Dict[str, str]) -> List[Instance]:
-        """
-        Helper for generating instances for a split.
-        Args:
-            splits (dict): Which splits to partition the data into.
-        Returns:
-            List[Instance]: Instances from the file for the specified split.
-        """
-        instances: List[Instance] = []
-        dataset: Any = load_dataset(
-            self.dataset_name,
-            self.subset,
-            revision=self.revision,
-            trust_remote_code=True,
-        )
-
-        for dataset_split_name, helm_split_name in splits.items():
-            for sample in dataset[dataset_split_name]:
-                target_sentence = sample[self.text_key]
-                source_sentence = self.__mask_text__(target_sentence)
-                instances.append(
-                    Instance(
-                        input=Input(text=source_sentence),
-                        references=[Reference(Output(text=target_sentence), tags=[CORRECT_TAG])],
-                        split=helm_split_name,
-                    )
-                )
-        return instances
-
-    def get_instances(self, output_path: str) -> List[Instance]:
-        if self.splits is None:
-            splits = {"train": TRAIN_SPLIT, "validation": VALID_SPLIT, "test": TEST_SPLIT}
-        else:
-            splits = {}
-            if "train" in self.splits:
-                splits[self.splits[TRAIN_SPLIT]] = TRAIN_SPLIT
-            if "validation" in self.splits:
-                splits[self.splits[VALID_SPLIT]] = VALID_SPLIT
-            if "test" in self.splits:
-                splits[self.splits[TEST_SPLIT]] = TEST_SPLIT
-
-        instances: List[Instance] = self.get_instances_for_splits(splits=splits)
-        return instances
-
-
-class MELTLMMaskFillingMLQAScenario(MELTLMMaskFillingScenario):
-    """
-    Scenario for the MLQA dataset.
-    This dataset is a multilingual question answering dataset.
-    It contains questions in multiple languages and their corresponding
-    answers in the same language. In this scenario, we are using the
-    context of questions in the Vietnamese subset of the MLQA dataset.
-    """
-
-    name = "melt_lm_mask_filling_mlqa"
-    description = "MLQA dataset for masked language modeling."
-    tags = ["language_modeling", "mask_filling"]
-
-    def __init__(self):
-        super().__init__(
-            dataset_name="facebook/mlqa",
-            revision="397ed406c1a7902140303e7faf60fff35b58d285",
-            subset="mlqa.vi.vi",
-            text_key="context",
-            splits={
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            },
-        )
-
-
-class MELTLMSpellingCorrectionScenario(Scenario):
-    """
-    Scenario for the MELT spelling correction dataset.
-    """
-
-    name = "melt_lm_spelling_correction"
-    description = "Spelling Correction scenario."
-    tags = ["language_modeling", "spelling_correction"]
-
-    def __init__(
-        self,
-        dataset_name: str,
-        revision: str,
-        source_key: str = "text",
-        target_key: str = "corrected_text",
-        subset: Optional[str] = "",
-        splits: Optional[Dict[str, str]] = None,
-    ):
-        """Initializes the question answering scenario.
-
-        Args:
-            dataset_name: The name of the dataset.
-            revision: The revision of the dataset to use.
-            source_key: The key to use for the source text in the dataset. Defaults to "text".
-            target_key: The key to use for the target text in the dataset. Defaults to "corrected_text".
-            subset: The subset of the dataset to use. Defaults to "".
-            splits: The splits to use for the dataset. Defaults to None.
-        """
-        super().__init__()
-        self.dataset_name = dataset_name
-        self.subset = subset
-        self.source_key = source_key
-        self.target_key = target_key
-        self.revision = revision
-        self.splits = splits
-
-    def get_instances_for_splits(self, splits: Dict[str, str]) -> List[Instance]:
-        """
-        Helper for generating instances for a split.
-        Args:
-            splits (dict): Which splits to partition the data into.
-        Returns:
-            List[Instance]: Instances from the file for the specified split.
-        """
-        instances: List[Instance] = []
-        dataset: Any = load_dataset(
-            self.dataset_name,
-            self.subset,
-            revision=self.revision,
-            trust_remote_code=True,
-        )
-        if len(splits) == 1:
-            all_keys = list(splits.keys())
-            dataset = dataset[all_keys[0]].train_test_split(test_size=0.33)
-            splits = {
-                "train": TRAIN_SPLIT,
-                "test": TEST_SPLIT,
-            }
-
-        for dataset_split_name, helm_split_name in splits.items():
-            for sample in dataset[dataset_split_name]:
-                source_sentence = sample[self.source_key]
-                target_sentence = sample[self.target_key]
-                instances.append(
-                    Instance(
-                        input=Input(text=source_sentence),
-                        references=[Reference(Output(text=target_sentence), tags=[CORRECT_TAG])],
-                        split=helm_split_name,
-                    )
-                )
-        return instances
-
-    def get_instances(self, output_path: str) -> List[Instance]:
-        if self.splits is None:
-            splits = {"train": TRAIN_SPLIT, "validation": VALID_SPLIT, "test": TEST_SPLIT}
-        else:
-            splits = {}
-            if "train" in self.splits:
-                splits[self.splits[TRAIN_SPLIT]] = TRAIN_SPLIT
-            if "validation" in self.splits:
-                splits[self.splits[VALID_SPLIT]] = VALID_SPLIT
-            if "test" in self.splits:
-                splits[self.splits[TEST_SPLIT]] = TEST_SPLIT
-
-        instances: List[Instance] = self.get_instances_for_splits(splits=splits)
-        return instances
-
-
-class MELTLMSpellingCorrectionVSECScenario(MELTLMSpellingCorrectionScenario):
-    """
-    Scenario for the VSEC dataset.
-    The VSEC dataset is a Vietnamese spelling correction dataset.
-    It contains 9,341 pairs of sentences where the first sentence is a misspelled
-    version of the second sentence, which is the correct version.
-    The mistakes are common spelling errors made by Vietnamese speakers and typists.
-    """
-
-    name = "melt_lm_spelling_correction_vsec"
-    description = "VSEC dataset for spelling correction."
-    tags = ["language_modeling", "spelling_correction"]
-
-    def __init__(self):
-        super().__init__(
-            dataset_name="ura-hcmut/VSEC",
-            revision="a6732e131605b5ec24ecc1745c6061c5ae86814e",
-            source_key="text",
-            target_key="correct",
-            splits={
-                TEST_SPLIT: "test",
-            },
-        )
-
-
-class MELTClosedBookQAScenario(Scenario):
-    name = "melt_closed_book_qa"
-    description = "Closed Book Question Answering scenario."
-    tags = ["question_answering"]
-
-    def __init__(
-        self,
-        dataset_name: str,
-        revision: str,
-        subset: Optional[str] = "",
-        splits: Optional[Dict[str, str]] = None,
-    ):
-        """
-        Initializes the question answering scenario.
-
-        Args:
-            dataset_name: The name of the dataset.
-            revision: The revision of the dataset to use.
-            subset: The subset of the dataset to use. Defaults to "".
-            splits: The splits to use for the dataset. Defaults to None.
-        """
-        super().__init__()
-        self.dataset_name = dataset_name
-        self.subset = subset
-        self.revision = revision
-        self.splits = splits
-
-    def get_instances_for_splits(self, splits: Dict[str, str]) -> List[Instance]:
-        """
-        Helper for generating instances for a split.
-        Args:
-            splits (dict): Which splits to partition the data into.
-        Returns:
-            List[Instance]: Instances from the file for the specified split.
-        """
-        instances: List[Instance] = []
-        dataset = load_dataset(
-            self.dataset_name,
-            self.subset,
-            revision=self.revision,
-            trust_remote_code=True,
-        )
-        for dataset_split_name, helm_split_name in splits.items():
-
-            for sample in dataset[dataset_split_name]:
-                instance = Instance(
-                    input=Input(text=sample["question"]),
-                    references=[Reference(Output(text=sample["answer"]), tags=[CORRECT_TAG])],
-                    split=helm_split_name,
-                )
-                instances.append(instance)
-
-        return instances
-
-    def get_instances(self, output_path: str) -> List[Instance]:
-        if self.splits is None:
-            splits = {"train": TRAIN_SPLIT, "validation": VALID_SPLIT, "test": TEST_SPLIT}
-        else:
-            splits = {}
-            if "train" in self.splits:
-                splits[self.splits[TRAIN_SPLIT]] = TRAIN_SPLIT
-            if "validation" in self.splits:
-                splits[self.splits[VALID_SPLIT]] = VALID_SPLIT
-            if "test" in self.splits:
-                splits[self.splits[TEST_SPLIT]] = TEST_SPLIT
-
-        instances: List[Instance] = self.get_instances_for_splits(splits=splits)
-        return instances
-
-
-class MELTKnowledgeZaloScenario(MELTClosedBookQAScenario):
-    """
-    Scenario for the Zalo dataset.
-    """
-
-    name = "melt_knowledge_zalo"
-    description = "Zalo dataset for closed-book question answering."
-    tags = ["question_answering", "knowledge"]
-
-    def __init__(self):
-        super().__init__(
-            dataset_name="ura-hcmut/zalo_e2eqa",
-            revision="63494521f4de949bfa57a5f0b79bc3ee47e635ad",
-            splits={
-                TRAIN_SPLIT: "train",
-                TEST_SPLIT: "test",
-            },
-        )
-
-
-class MELTMultipleChoiceQAScenario(Scenario):
-    name = "melt_multiple_choice_qa"
-    description = "Multiple Choice Question Answering scenario."
-    tags = ["question_answering"]
-
-    def __init__(
-        self,
-        dataset_name: str,
-        revision: str,
-        subset: Optional[str] = "",
-        splits: Optional[Dict[str, str]] = None,
-    ):
-        """
-        Initializes the question answering scenario.
-
-        Args:
-            dataset_name: The name of the dataset.
-            revision: The revision of the dataset to use.
-            subset: The subset of the dataset to use. Defaults to "".
-            splits: The splits to use for the dataset. Defaults to None.
-        """
-        super().__init__()
-        self.dataset_name = dataset_name
-        self.subset = subset
-        self.revision = revision
-        self.splits = splits
-
-    @abstractmethod
-    def process_example(self, sample: dict) -> Tuple[Input, List[Reference]]:
-        """
-        Given an sample from the dataset, create the input text and
-        list of answers for the instance.
-        """
-        pass
-
-    def get_instances_for_splits(self, splits: Dict[str, str]) -> List[Instance]:
-        """
-        Helper for generating instances for a split.
-        Args:
-            splits (dict): Which splits to partition the data into.
-        Returns:
-            List[Instance]: Instances from the file for the specified split.
-        """
-        instances: List[Instance] = []
-        dataset = load_dataset(
-            self.dataset_name,
-            self.subset,
-            revision=self.revision,
-            trust_remote_code=True,
-        )
-        for dataset_split_name, helm_split_name in splits.items():
-            for sample in dataset[dataset_split_name]:
-                inputs, references = self.process_example(sample)
-                instance = Instance(
-                    # input=sample["question"],
-                    # references=[Reference(Output(text=sample["answer"]), tags=[CORRECT_TAG])],
-                    input=inputs,
-                    references=references,
-                    split=helm_split_name,
-                )
-                instances.append(instance)
-
-        return instances
-
-    def get_instances(self, output_path: str) -> List[Instance]:
-        if self.splits is None:
-            splits = {"train": TRAIN_SPLIT, "validation": VALID_SPLIT, "test": TEST_SPLIT}
-        else:
-            splits = {}
-            if "train" in self.splits:
-                splits[self.splits[TRAIN_SPLIT]] = TRAIN_SPLIT
-            if "validation" in self.splits:
-                splits[self.splits[VALID_SPLIT]] = VALID_SPLIT
-            if "test" in self.splits:
-                splits[self.splits[TEST_SPLIT]] = TEST_SPLIT
-
-        instances: List[Instance] = self.get_instances_for_splits(splits=splits)
-        return instances
-
-
-class MELTKnowledgeViMMRCScenario(MELTMultipleChoiceQAScenario):
-    """
-    Scenario for the ViMMRC dataset.
-    """
-
-    name = "melt_knowledge_vimmrc"
-    description = "ViMMRC dataset for multiple choice question answering."
-    tags = ["question_answering", "knowledge"]
-
-    def __init__(self, randomize_order: bool = False):
-        super().__init__(
-            dataset_name="ura-hcmut/ViMMRC",
-            revision="fe68800e37aaa84d80b1d93466b36c3fa60d8bcb",
-            splits={
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            },
-        )
-        self.randomize_order = randomize_order
-        self.correct_answer_mapping = {
-            "A": 0,
-            "B": 1,
-            "C": 2,
-            "D": 3,
-            "E": 4,
-            "F": 5,
-            "G": 6,
-            "H": 7,
-        }
-        random.seed(42)
-
-    def process_example(self, sample: dict) -> Tuple[Input, List[Reference]]:
-        """
-        Given an sample from the dataset, create the input text and
-        list of answers for the instance.
-        """
-        inputs = PassageQuestionInput(
-            passage=sample["article"],
-            passage_prefix="Ngữ cảnh: ",
-            question=sample["question"],
-            question_prefix="Câu hỏi: ",
-            separator="\n\n",
-        )
-
-        correct_idx = self.correct_answer_mapping[sample["answer"]]
-        references = []
-        for idx, answer in enumerate(eval(sample["options"])):
-            if idx == correct_idx:
-                tags = [CORRECT_TAG]
-            else:
-                tags = []
-
-            references.append(Reference(Output(text=answer), tags=tags))
-
-        if self.randomize_order:
-            k = random.randint(0, len(references) - 1)
-            references = references[k:] + references[:k]
-
-        return inputs, references
-
-
 class MELTToxicityDetectionViHSDScenario(MELTTextClassificationScenario):
     """
     Scenario for the UIT-ViHSD dataset.
@@ -1407,157 +658,3 @@ class MELTToxicityDetectionViCTSDScenario(MELTTextClassificationScenario):
         """
         label = sample[self.label_key]
         return sample[self.text_key], [self.label_mapping[label]]
-
-
-class MELTInformationRetrievalScenario(Scenario):
-    name = "melt_information_retrieval"
-    description = "Scenario for information retrieval tasks."
-    tags = ["information_retrieval"]
-
-    """ Dictionary mapping task track tuples to the number of queries. """
-    NUM_TRAIN_QUERIES = 1000
-
-    """ Upper and lower bounds on top-k.
-
-    The top-k number represents the number of passages we will consider per
-    query. Max top-k for the train and validation files are set to the number
-    of passages included in the corresponding top-k files.
-    """
-    MIN_TOPK: int = 11
-    MAX_TRAIN_TOPK: int = 20
-    MAX_VALID_TOPK: int = 1000
-
-    def __init__(self, dataset_name: str, revision: str, subset: Optional[str] = "", valid_topk: Optional[int] = None):
-        """The constructor for the MSMARCOScenario.
-
-        Args:
-            dataset_name: The name of the dataset.
-            revision: The revision of the dataset to use.
-            subset: The subset of the dataset to use. Defaults to "".
-            valid_topk: If set, specifies the number of top documents for which the
-                validation instances will be created. Must be in the range
-                [self.MIN_TOPK, self.MAX_VALID_TOPK].
-        """
-        super().__init__()
-
-        # Input validation
-        self.dataset_name = dataset_name
-        self.revision = revision
-        self.subset = subset
-        self.valid_topk: Optional[int] = valid_topk
-        if self.valid_topk is not None:
-            assert valid_topk and self.MIN_TOPK <= valid_topk <= self.MAX_VALID_TOPK
-
-    def get_train_instances(self) -> List[Instance]:
-        """Get training instances.
-        References for each instance are selected as follows:
-            1. We select 1 correct reference, where the documents included
-            corresponds to the best document for the given train query.
-            2. We create 1 wrong reference, where the document included
-            corresponds to a non-gold document for the given train query.
-        """
-        dataset = load_dataset(
-            self.dataset_name,
-            self.subset,
-            revision=self.revision,
-            trust_remote_code=True,
-        )
-        instances = []
-        for i, sample in enumerate(dataset["train"]):
-
-            if i >= self.NUM_TRAIN_QUERIES:
-                break
-
-            references = [
-                Reference(Output(text=sample["positive"]), tags=[CORRECT_TAG]),
-                Reference(Output(text=sample["negative"]), tags=[]),
-            ]
-
-            instances.append(Instance(Input(text=sample["query"]), references=references, split=TRAIN_SPLIT))
-        return instances
-
-    def get_valid_instances(self) -> List[Instance]:
-        """Get validation instances.
-        By default, we create a reference for each Document ID for which there
-        is a judgment with respect to the provided Query ID.
-
-        If self.valid_topk is not None, we ensure that a reference is created
-        for all the documents that appear in top self.valid_topk documents for
-        the given validation query.
-        """
-        dataset = load_dataset(
-            self.dataset_name,
-            f"runs-{self.subset}",
-            revision=self.revision,
-            trust_remote_code=True,
-        )
-        instances = []
-        for sample in dataset["bm25"]:
-            references = []
-
-            for k, passage_dict in enumerate(Dataset.from_dict(sample["passages"])):
-                if self.valid_topk is None or k >= self.valid_topk:
-                    break
-                tags = []
-                tags.append(f"docid={passage_dict['id']}")
-                if k == 0:
-                    tags.append(CORRECT_TAG)
-                tags.append(make_rank_tag(rank=k + 1))  # Top-k rank
-                references.append(Reference(Output(text=passage_dict["passage"]), tags=tags))
-
-            instances.append(Instance(Input(text=sample["query"]), references=references, split=VALID_SPLIT))
-
-        return instances
-
-    def get_instances(self, output_path: str) -> List[Instance]:
-        """Get instances for this scenario.
-
-        Refer to the documentation of the following methods for details on how
-        the instances are created:
-            * self.get_train_instances
-            * self.get_valid_instances
-        """
-
-        hlog("Preparing training instances.")
-        train_instances = self.get_train_instances()
-
-        hlog("Preparing validation instances.")
-        valid_instances = self.get_valid_instances()
-
-        return train_instances + valid_instances
-
-
-class MELTInformationRetrievalMMARCOScenario(MELTInformationRetrievalScenario):
-    """
-    Scenario for the MMARCO dataset.
-    """
-
-    name = "melt_information_retrieval_mmarco"
-    description = "MMARCO dataset for information retrieval in Vietnamese."
-    tags = ["information_retrieval"]
-
-    def __init__(self, **kwargs):
-        super().__init__(
-            dataset_name="unicamp-dl/mmarco",
-            revision="6d039c4638c0ba3e46a9cb7b498b145e7edc6230",
-            subset="vietnamese",
-            **kwargs,
-        )
-
-
-class MELTInformationRetrievalMRobustScenario(MELTInformationRetrievalScenario):
-    """
-    Scenario for the MRobust dataset.
-    """
-
-    name = "melt_information_retrieval_mrobust"
-    description = "MRobust dataset for information retrieval in Vietnamese."
-    tags = ["information_retrieval"]
-
-    def __init__(self, **kwargs):
-        super().__init__(
-            dataset_name="unicamp-dl/mrobust",
-            revision="fda452a7fbfd9550db2f78d9d98e6b3ec16734df",
-            subset="vietnamese",
-            **kwargs,
-        )
